backend/extract/g_fit/requirements.txt
google-auth >= 1.23.0
# google-auth-oauthlib >= 0.8.0
# google-auth-httplib2 >= 0.0.1
google-api-python-client >= 2.76.0
requests >= 2.18.4
flask >= 2.2.2
google.cloud.functions >= 1.10.1
functions-framework >= 3.3.0
requests-oauthlib >= 1.3.1